update Layout by removing DefaultLayout
